+++ packages/ketcher-react/src/script/ui/views/components/Spinner/Spinner.module.less
