+++ packages/ketcher-react/src/script/ui/views/components/Spinner/index.ts
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2020 EPAM Systems
+ * Copyright 2021 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,4 @@
  * limitations under the License.
  ***************************************************************************/
 
-import Spinner from './Spinner'
-
-export { Spinner }
+export { default } from './Spinner'
