+++ packages/ketcher-react/src/script/ui/views/components/StructEditor/StructEditor.jsx
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2020 EPAM Systems
+ * Copyright 2021 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,7 @@
 import { upperFirst } from 'lodash/fp'
 import React, { Component, createRef } from 'react'
 import Editor from '../../../../editor'
-import { Spinner } from './components'
+import Spinner from '../Spinner'
 
 import styles from './StructEditor.module.less'
