+++ packages/ketcher-react/src/script/ui/views/components/index.ts
@@ -1,5 +1,5 @@
 /****************************************************************************
- * Copyright 2020 EPAM Systems
+ * Copyright 2021 EPAM Systems
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,6 +16,6 @@
 
 import Dialog from './Dialog'
 import StructEditor from './StructEditor'
-import Spinner from './StructEditor/components/Spinner'
+import Spinner from './Spinner'
 
 export { Dialog, StructEditor, Spinner }
